Tidy up environment object
