+++ src/Adapters/Laravel/Facades/Phiki.php
@@ -5,7 +5,7 @@ namespace Phiki\Adapters\Laravel\Facades;
 use Illuminate\Support\Facades\Facade;
 
 /**
- * @method static \Phiki\Environment\Environment environment()
+ * @method static \Phiki\Environment environment()
  * @method static array<int, array<int, \Phiki\Token\Token>> codeToTokens(string $code, string|\Phiki\Grammar\Grammar|\Phiki\Grammar\ParsedGrammar $grammar)
  * @method static array<int, array<int, \Phiki\Token\HighlightedToken>> tokensToHighlightedTokens(array<int, array<int, \Phiki\Token\Token>> $tokens, string|array|\Phiki\Theme\Theme $theme)
  * @method static array<int, array<int, \Phiki\Token\HighlightedToken>> codeToHighlightedTokens(string $code, string|\Phiki\Grammar\Grammar $grammar, string|array|\Phiki\Theme\Theme $theme)
