+++ src/TextMate/Tokenizer.php
@@ -2,7 +2,7 @@
 
 namespace Phiki\TextMate;
 
-use Phiki\Environment\Environment;
+use Phiki\Environment;
 use Phiki\Grammar\BeginEndPattern;
 use Phiki\Grammar\BeginWhilePattern;
 use Phiki\Grammar\EndPattern;
@@ -311,7 +311,7 @@ class Tokenizer
      */
     protected function matchRule(ParsedGrammar $grammar, string $lineText, bool $isFirstLine, int $linePos, StateStack &$stack, int $anchorPosition): ?MatchedPattern
     {
-        return (new PatternSearcher($stack->pattern, $grammar, $this->environment->getGrammarRepository(), $isFirstLine, $linePos === $anchorPosition))
+        return (new PatternSearcher($stack->pattern, $grammar, $this->environment->grammars, $isFirstLine, $linePos === $anchorPosition))
             ->findNextMatch($lineText, $linePos);
     }
 
@@ -338,7 +338,7 @@ class Tokenizer
                 continue;
             }
 
-            $searcher = new PatternSearcher($injection, $grammar, $this->environment->getGrammarRepository(), $isFirstLine, $linePos === $anchorPosition);
+            $searcher = new PatternSearcher($injection, $grammar, $this->environment->grammars, $isFirstLine, $linePos === $anchorPosition);
             $matched = $searcher->findNextMatch($lineText, $linePos);
 
             if (! $matched) {
@@ -384,7 +384,7 @@ class Tokenizer
 
         // Process and check each while pattern.
         for ($whileRule = array_pop($whileRules); $whileRule; $whileRule = array_pop($whileRules)) {
-            $searcher = new PatternSearcher($whileRule->rule, $grammar, $this->environment->getGrammarRepository(), $isFirstLine, $linePos === $anchorPosition);
+            $searcher = new PatternSearcher($whileRule->rule, $grammar, $this->environment->grammars, $isFirstLine, $linePos === $anchorPosition);
             $r = $searcher->findNextMatch($lineText, $linePos, while: true);
 
             if (! $r) {
